skip empty registries, add doc
